pipeline/pipeline/requirements.txt
# 核心Web框架
Flask==3.0.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Werkzeug==3.0.1

# Google AI依赖 (支持多模型切换)
google-generativeai==0.8.3

# 图像处理 (支持多图片上传)
Pillow==10.4.0

# 语音处理
SpeechRecognition==3.10.4
gTTS==2.5.1
pydub==0.25.1

# WebSocket支持 (实时通信)
websockets==12.0
python-socketio==5.11.0

# HTTP请求 (API调用和转发)
requests==2.31.0
urllib3==2.1.0

# ROS2支持 (可选，用于机器人通信)
# rclpy>=3.0.0  # 取消注释以启用ROS2支持
# std_msgs>=4.0.0

# 音频处理
pyaudio==0.2.14

# 日志和配置
python-dotenv==1.0.1
colorama==0.4.6

# 数据处理
numpy>=1.24.0
json5==0.9.14

# 系统工具
psutil==5.9.8

# 可选：音频格式转换
ffmpeg-python==0.2.0

# 开发和测试工具 (可选)
# pytest==7.4.4
# pytest-asyncio==0.23.2

# 安全相关
cryptography>=41.0.0

# 异步支持 (已内置在Python 3.7+中，无需单独安装)
# asyncio 已移除，使用内置版本
